extract common logic into new AbstractConnectionResolver parent class
