+++ core/domain/services/graphql/connections/RootQueryDatetimesConnection.php
@@ -2,29 +2,20 @@
 
 namespace EventEspresso\core\domain\services\graphql\connections;
 
-use EE_Base_Class;
 use EEM_Datetime;
 use EventEspresso\core\domain\services\graphql\connection_resolvers\DatetimeConnectionResolver;
-use EventEspresso\core\domain\services\graphql\connections\EventDatetimesConnection;
-use EventEspresso\core\services\graphql\connections\ConnectionInterface;
 use Exception;
-use WPGraphQL\Type\WPObjectType;
 
 /**
  * Class RootQueryDatetimesConnection
  * Description
  *
  * @package EventEspresso\core\domain\services\graphql\connections
- * @author  Brent Christensen
+ * @author  Manzoor Ahmad Wani
  * @since   $VID:$
  */
-class RootQueryDatetimesConnection implements ConnectionInterface
+class RootQueryDatetimesConnection extends AbstractRootQueryConnection
 {
-
-    /**
-     * @var EEM_Datetime $model
-     */
-    protected $model;
 
 
     /**
@@ -61,27 +52,12 @@ class RootQueryDatetimesConnection implements ConnectionInterface
      * @param $args
      * @param $context
      * @param $info
-     * @return array
+     * @return DatetimeConnectionResolver
      * @throws Exception
      * @since $VID:$
      */
-    public function resolveConnection($entity, $args, $context, $info)
+    public function getConnectionResolver($entity, $args, $context, $info)
     {
-        $resolver = new DatetimeConnectionResolver($entity, $args, $context, $info);
-        return $resolver->get_connection();
-    }
-
-
-    /**
-     * @param $id
-     * @param $args
-     * @param $context
-     * @param $info
-     * @return EE_Base_Class
-     * @since $VID:$
-     */
-    public function resolveNode($id, $args, $context, $info)
-    {
-        return $this->model->get_one_by_ID($id);
+        return new DatetimeConnectionResolver($entity, $args, $context, $info);
     }
 }
