+++ core/domain/services/graphql/connections/RootQueryTicketsConnection.php
@@ -2,23 +2,19 @@
 
 namespace EventEspresso\core\domain\services\graphql\connections;
 
-use EE_Base_Class;
 use EEM_Ticket;
 use EventEspresso\core\domain\services\graphql\connection_resolvers\TicketConnectionResolver;
-use EventEspresso\core\domain\services\graphql\connections\DatetimeTicketsConnection;
-use EventEspresso\core\services\graphql\connections\ConnectionInterface;
 use Exception;
-use WPGraphQL\Type\WPObjectType;
 
 /**
  * Class RootQueryTicketsConnection
  * Description
  *
  * @package EventEspresso\core\domain\services\graphql\connections
- * @author  Brent Christensen
+ * @author  Manzoor Ahmad Wani
  * @since   $VID:$
  */
-class RootQueryTicketsConnection implements ConnectionInterface
+class RootQueryTicketsConnection extends AbstractRootQueryConnection
 {
 
     /**
@@ -61,27 +57,12 @@ class RootQueryTicketsConnection implements ConnectionInterface
      * @param $args
      * @param $context
      * @param $info
-     * @return array
+     * @return TicketConnectionResolver
      * @throws Exception
      * @since $VID:$
      */
-    public function resolveConnection($entity, $args, $context, $info)
+    public function getConnectionResolver($entity, $args, $context, $info)
     {
-        $resolver = new TicketConnectionResolver($entity, $args, $context, $info);
-        return $resolver->get_connection();
-    }
-
-
-    /**
-     * @param $id
-     * @param $args
-     * @param $context
-     * @param $info
-     * @return EE_Base_Class
-     * @since $VID:$
-     */
-    public function resolveNode($id, $args, $context, $info)
-    {
-        return $this->model->get_one_by_ID($id);
+        return new TicketConnectionResolver($entity, $args, $context, $info);
     }
 }
